+++ docs/src/shared/components/ReactMD/data-tables/PaginationExample/code.js
@@ -1,0 +1,12 @@
+import PaginationExample from '!!raw!./index';
+import Loader from '!!raw!./Loader';
+
+export default `
+/* PaginationExample.jsx */
+${PaginationExample}
+\`\`\`
+
+\`\`\`js
+/* Loader.jsx */
+${Loader}
+`;