+++ docs/src/shared/components/ReactMD/data-tables/index.js
@@ -3,84 +3,55 @@ import React from 'react';
 import PlainTableExample from './PlainTableExample';
 import PlainTableExampleRaw from '!!raw!./PlainTableExample';
 
-import DataTableExample from './DataTableExample';
-import DataTableExampleRaw from '!!raw!./DataTableExample';
+import DynamicContentTable from './DynamicContentTable';
+import DynamicContentTableRaw from './DynamicContentTable/code';
 
-import TablesInCards from './TablesInCards';
-import TablesInCardsRaw from '!!raw!./TablesInCards';
-
-import ComplexDataTableComments from './ComplexDataTableComments';
-import ComplexDataTableCommentsRaw from '!!raw!./ComplexDataTableComments';
+import ConfigurableTableExample from './ConfigurableTableExample';
+import ConfigurableTableExampleRaw from './ConfigurableTableExample/code';
 
 import PaginationExample from './PaginationExample';
-import PaginationExampleRaw from '!!raw!./PaginationExample';
-import PaginationLoaderRaw from '!!raw!./PaginationLoader';
-
-import ControlledPagination from './ControlledPagination';
-import ControlledPaginationRaw from '!!raw!./ControlledPagination';
+import PaginationExampleRaw from './PaginationExample/code';
 
 export default [{
   title: 'Plain Table Example',
+  description: `
+A plain table removes the checkboxes from the table and updates the cells to not
+enforce a strict height. This allows for more dynamic and multiple line content.
+  `,
   code: PlainTableExampleRaw,
   children: <PlainTableExample />,
   tableCard: true,
 }, {
-  title: 'Data Table Example',
+  title: 'Dynamic Content Table',
   description: `
-When the \`plain\` prop is not enabled, the \`DataTable\` will inject a checkbox at the start of each
-row which will allow the user to select that row.
+This example will show how you can make a fully interactive data table that allows a user
+to insert/remove rows. This exmaple will use the \`EditDialogColumn\` for the inline cell edits
+and the \`SelectFieldColumn\` for the select fields.
 
-This example also introduces another component: \`SelectFieldColumn\`. When using the \`SelectFieldColumn\`,
-your \`TableColumn\` in the \`TableHeader\` should apply the \`.md-table-column--select-field\` to position
-the header with your select field. In addition, it is recommended to specify an exact width or min-width for
-your column on your column's header. When the values for the select field are different widths, the column's
-size will keep changing.
-`,
-  code: DataTableExampleRaw,
-  children: <DataTableExample />,
+When tables are placed in cards, it can be useful to have additional interactions such as inserting
+and removing rows. The \`TableCardHeader\` is the component that will manage this for you. This component
+will display as a \`CardTitle\` until the \`visible\` prop is enabled. It will then show a "contextual header"
+that will show some additional actions or a different title. This is normally used when rows in the table
+have been selected.
+  `,
+  code: DynamicContentTableRaw,
+  children: <DynamicContentTable />,
+}, {
+  title: 'Configurable Data Table Example',
+  description: `
+This example will show some of the different configuration options for the \`TableColumn\` and \`EditDialogColumn\`.
+The \`TableColumn\` can be used to show a sort icon based on the \`sorted\` prop. The sorting will have to happen
+manually on your end.
+
+The \`EditDialogColumn\` can be displayed in a dialog, a large dialog, or inline. The defalt behavior for the dialog
+versions is to "save" the cell when the user presses enter. If the user hits escape, their changes will be undone and
+the value will be reset before the dialog was opened.
+  `,
+  code: ConfigurableTableExampleRaw,
+  children: <ConfigurableTableExample />,
   tableCard: true,
 }, {
-  title: 'Tables In Cards',
-  description: `
-When using \`DataTable\`s in \`Card\`s, it might be required that additional actions or contextual
-events happen when the user selects rows. There is a \`TableCardHeader\` component that can be used
-to help with this.
-
-The \`TableCardHeader\` can either display a \`title\` with \`children\` OR \`leftChildren\` with \`children\`.
-The \`children\` prop _should_ normally be some icon buttons or \`MenuButton\` to do some additional actions.
-By default, the first child will have the \`.md-cell--right\` class name cloned into it to force them to the
-far right of the header. This can be disabled by providing \`noChildrenAdjust\`.
-
-When the \`leftChildren\` prop is defined, it assumes that it is either an array of flat buttons, or a single
-flat button. It will clone the \`.md-btn--dialog\` className into each button to style some spacing and change
-the min-width. This behavior can be disabled by providing \`noLeftChildrenClone\`.
-
-Once the user has selected a row or multiple rows, a contextual header will appear that will display the count
-of selected rows as well as some new \`actions\` that can be achieved.
-`,
-  code: TablesInCardsRaw,
-  children: <TablesInCards />,
-}, {
-  title: 'Complex Table Example',
-  description: `
-When using an edit dialog in the \`DataTable\`, the default behavior is to prevent the user to tab away
-once they have focused a dialog. The user must either press \`enter\` or \`esc\` to save or cancel their
-edit respectively.
-
-If the \`EditDialog\` is \`large\`, the user's focus will be contained between the text field, the cancel button,
-and the save button. In all cases, the user can click away from the dialog to close it. The text can either
-be saved or cancelled when this happens by using the \`okOnOutsideClick\` prop.
-
-TableColumns can also inject an icon for indicating a column is sortable. When the \`sorted\` prop is true,
-the column will be updated with \`aria-sort="ascending"\` and if the \`sorted\` prop is false, it will be
-updated with \`aria-sort="descending"\`. Only one column should be sortable at a time according to
-[W3C Recommendations](https://www.w3.org/TR/wai-aria/states_and_properties#aria-sort).
-`,
-  code: ComplexDataTableCommentsRaw,
-  children: <ComplexDataTableComments />,
-  tableCard: true,
-}, {
-  title: 'Table Pagination Example',
+  title: 'Pagination Example',
   description: `
 When working with giant datasets, you can of course increase performance by limiting the number of rows visible
 at a time. This can be accomplished by the \`TablePagination\` component.
@@ -92,27 +63,7 @@ and buttons to either return to the previous page or advance.
 
 It will be up to you to splice your dataset correctly from this component.
 `,
-  code: `
-/* PaginationExample.jsx */
-${PaginationExampleRaw}
-\`\`\`
-
-\`\`\`js
-/* PaginationLoader.jsx */
-${PaginationLoaderRaw}
-`,
+  code: PaginationExampleRaw,
   children: <PaginationExample />,
-  tableCard: true,
-}, {
-  title: 'Controlled Pagination Example',
-  description: `
-The \`TablePagination\` component can also be fully controlled by passing a \`page\` and \`rowsPerPage\` prop. This allows you to use query parameters or
-anything else to handle the pagination if that's your thing. Unfortunately, I'm too lazy to implement that in this website so I am going to show an internal
-state/controlled example.
-
-It is probably still more helpful just to use the \`defaultPage\` and \`defaultRowsPerPage\` props.
-  `,
-  code: ControlledPaginationRaw,
-  children: <ControlledPagination />,
   tableCard: true,
 }];