+++ src/js/SelectionControls/SelectionControlGroup.js
@@ -180,6 +180,11 @@ export default class SelectionControlGroup extends PureComponent {
      * Boolean if the `SelectionControl` should be displayed inline.
      */
     inline: PropTypes.bool,
+
+    /**
+     * Boolean if all the selection controls in the group are disabled.
+     */
+    disabled: PropTypes.bool,
   };
 
   static defaultProps = {
@@ -245,6 +250,7 @@ export default class SelectionControlGroup extends PureComponent {
       component: Component,
       labelComponent: LabelComponent,
       inline,
+      disabled,
       /* eslint-disable no-unused-vars */
       value: propValue,
       controls: propControls,
@@ -262,6 +268,7 @@ export default class SelectionControlGroup extends PureComponent {
         name: `${name}${type === 'checkbox' ? '[]' : ''}`,
         type,
         inline,
+        disabled,
         checked: this._isChecked(value, control.value, type),
       }, control, {
         style: Object.assign({}, control.style, controlStyle),
@@ -285,7 +292,6 @@ export default class SelectionControlGroup extends PureComponent {
         {ariaLabel}
         {controls}
       </Component>
-
     );
   }
 }